bots/music_bot/requirements.txt
aiogram
yt-dlp
aiohttp
python-dotenv